Start of my item cosmetic :)
TL;DR every time you use a trapper bow, it stores the color of the bow in command storage. When tracking bomb is used it spawns dust particles of each of those colors in order behind it and spawns all of them as one dust particle when it explodes.

Current issues:
- Error when running macro function, array of floats for dust particle color is recognized as incorrect data type
- Short blast doesn't correctly store its color on use
- No implementation for particles when tracking bomb explodes
- No functionality for toggling the cosmetic

Committing while I wait for answers on that first issue on MCC.